[FEAT] 스플래시 기능 추가 - 앱 아이콘 - 텍스트 로고 - 스플래시 Theme - 패키징 ui/#16
