+++ app/src/main/AndroidManifest.xml
@@ -14,9 +14,9 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".ui.user.LoginActivity"
-            android:exported="true">
-
+            android:name=".presentation.splash.SplashActivity"
+            android:exported="true"
+            android:theme="@style/Theme.YeoGida.Splash">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -29,11 +29,24 @@
         </activity>
 
         <activity
+            android:name=".presentation.user.LoginActivity"
+            android:exported="true">
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
+
+        <activity
             android:name=".MainActivity"
             android:exported="true">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
+        </activity>
+
+        <activity
+            android:name=".presentation.user.JoinActivity"
+            android:exported="true">
         </activity>
 
     </application>
