+++ app/src/main/java/com/starters/yeogida/presentation/follow/FollowFragment.kt
@@ -1,10 +1,11 @@
-package com.starters.yeogida
+package com.starters.yeogida.presentation.follow
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.starters.yeogida.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
