+++ app/src/main/java/com/starters/yeogida/presentation/like/LikeFragment.kt
@@ -1,10 +1,11 @@
-package com.starters.yeogida
+package com.starters.yeogida.presentation.like
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.starters.yeogida.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -13,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [AroundFragment.newInstance] factory method to
+ * Use the [LikeFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class AroundFragment : Fragment() {
+class LikeFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -34,7 +35,7 @@ class AroundFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_around, container, false)
+        return inflater.inflate(R.layout.fragment_like, container, false)
     }
 
     companion object {
@@ -44,12 +45,12 @@ class AroundFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment AroundFragment.
+         * @return A new instance of fragment LikeFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            AroundFragment().apply {
+            LikeFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
