+++ app/src/main/java/com/starters/yeogida/presentation/user/LoginActivity.kt
@@ -1,7 +1,7 @@
-package com.starters.yeogida.ui.user
+package com.starters.yeogida.presentation.user
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 import com.starters.yeogida.R
 
 class LoginActivity : AppCompatActivity() {
