+++ app/src/main/res/drawable/ic_app.xml
@@ -1,0 +1,37 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="64dp"
+    android:height="64dp"
+    android:viewportWidth="1024"
+    android:viewportHeight="1024">
+  <path
+      android:pathData="M0,0h1024v1024h-1024z"
+      android:fillColor="#439AFF"/>
+  <group
+    android:scaleX="0.9"
+    android:scaleY="0.9"
+    android:pivotX="700"
+    android:pivotY="600">
+    <path
+        android:pathData="M660.5,215C631.7,215 603.3,221.7 577.5,234.6C551.7,247.5 529.3,266.2 512,289.3C488.6,258.1 456,235.1 418.8,223.4C381.7,211.8 341.8,212.2 304.8,224.5C267.8,236.8 235.7,260.5 212.9,292.1C190.1,323.7 177.9,361.7 177.9,400.6C177.9,623.1 486.9,798.5 499.9,805.9C503.6,808 507.8,809 512,809C516.2,809.1 520.4,808 524.1,805.9C581,772.6 634.4,733.7 683.4,689.6C791.4,592.4 846.1,495.3 846.1,400.6C846.1,351.4 826.6,304.2 791.8,269.4C756.9,234.6 709.7,215 660.5,215Z"
+        android:fillColor="#ffffff"/>
+  </group>
+
+  <path
+      android:pathData="M560,406.5a106,105.5 0,1 0,212 0a106,105.5 0,1 0,-212 0z"
+      android:fillColor="#FFBF43"/>
+  <path
+      android:pathData="M563.5,463C572.9,465 577.6,474.1 574.1,483.4L566.4,503.6C562.4,514.3 549.4,520.2 539.7,515.8L448.9,474.6C443.2,472 440,466.3 440.8,459.8C442.2,448.8 454,440.3 464.8,442.6L563.5,463Z"
+      android:fillColor="#439AFF"/>
+  <path
+      android:pathData="M583.1,492.8C581.3,483.4 572.4,478.5 563,481.8L542.6,489C531.9,492.8 525.6,505.6 529.8,515.4L569,607.1C571.5,612.9 577.2,616.2 583.7,615.5C594.7,614.4 603.4,602.8 601.4,591.9L583.1,492.8Z"
+      android:fillColor="#439AFF"/>
+  <path
+      android:pathData="M436,580.1C442,581.3 445,587.1 442.7,593.1L437.8,606C435.3,612.8 427,616.6 420.8,613.8L362.9,587.5C359.2,585.8 357.2,582.2 357.7,578C358.6,571 366.1,565.6 373,567L436,580.1Z"
+      android:fillColor="#439AFF"/>
+  <path
+      android:pathData="M447.7,596.5C448.2,590.4 443.5,585.9 437.2,586.4L423.4,587.5C416.2,588.1 410.2,594.9 411.2,601.7L420.2,664.6C420.8,668.6 423.7,671.6 427.8,672.3C434.8,673.4 442.1,667.7 442.7,660.6L447.7,596.5Z"
+      android:fillColor="#439AFF"/>
+  <path
+      android:pathData="M609,415.9C620.8,405.6 638.7,407.1 648.4,419.3C658.2,431.5 655.8,449.3 643.3,458.5L436.4,610.1C429.4,615.2 419.7,613.9 414.3,607.2C408.9,600.5 409.8,590.7 416.3,585L609,415.9Z"
+      android:fillColor="#439AFF"/>
+</vector>
