+++ app/src/main/res/drawable/img_text_logo.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="100dp"
+    android:height="40dp"
+    android:viewportWidth="141  "
+    android:viewportHeight="65">
+  <path
+      android:pathData="M23.91,5.96C24.07,5.54 24.38,5.3 24.82,5.24C25.3,5.18 25.66,5.35 25.88,5.77C26.17,6.02 25.96,7.29 25.26,9.56C24.71,11.32 24.12,13.46 23.48,15.99C23.06,17.66 22.63,19.22 22.18,20.7C21.9,21.69 21.75,22.42 21.75,22.9C21.75,23.32 21.38,24.39 20.65,26.12C19.94,27.85 19.29,28.71 18.68,28.71C18.17,28.71 17.9,28.46 17.86,27.94C17.86,27.43 18.1,26.58 18.58,25.4C18.97,24.47 19.5,22.66 20.17,19.98C20.87,17.29 21.22,15.67 21.22,15.13C21.22,14.26 19.5,14.49 16.04,15.8C14.15,16.5 12.84,16.92 12.1,17.05C11.4,17.18 11,17.08 10.9,16.76C10.74,16.34 11.66,15.69 13.64,14.79C15.62,13.86 17.42,13.27 19.02,13.02C20.9,12.66 21.93,12.39 22.09,12.2C22.18,12.07 22.31,11.7 22.47,11.1L22.81,9.99L17.77,10.28C14.34,10.47 12.33,10.62 11.72,10.71C11.05,10.81 10.66,10.57 10.57,9.99C10.5,9.64 10.6,9.35 10.86,9.13C11.11,8.9 11.53,8.73 12.1,8.6C12.71,8.44 13.5,8.33 14.46,8.26C15.45,8.17 16.68,8.12 18.15,8.12C19.53,8.12 20.58,8.1 21.32,8.07C22.06,8.04 22.58,7.98 22.9,7.88C23.22,7.78 23.42,7.62 23.48,7.4C23.51,7.3 23.56,7.16 23.62,6.97C23.69,6.78 23.74,6.58 23.77,6.39C23.83,6.2 23.88,6.06 23.91,5.96ZM3.27,10.71C4.78,10.01 5.9,10.78 6.63,13.02C7.3,15.29 6.95,17.1 5.58,18.44C5.03,18.98 4.22,19.3 3.13,19.4C2.04,19.5 1.26,19.3 0.78,18.82C0.04,18.12 -0.01,16.71 0.63,14.6C1.27,12.49 2.15,11.19 3.27,10.71ZM5.43,14.31C5.21,13.8 4.84,13.54 4.33,13.54C3.85,13.54 3.37,14.07 2.89,15.13C2.41,16.15 2.34,16.84 2.7,17.19C2.92,17.42 3.27,17.35 3.75,17C4.23,16.65 4.65,16.2 5,15.66C5.38,15.08 5.53,14.63 5.43,14.31Z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M70.27,0.87C70.63,0.46 70.96,0.31 71.28,0.44C71.76,0.63 71.86,1.5 71.57,3.03C71.31,4.57 70.75,6.36 69.89,8.41C68.87,10.78 68.07,12.94 67.49,14.89C66.95,16.81 66.26,19.69 65.43,23.53C64.5,28.14 63.73,30.44 63.12,30.44C62.99,30.44 62.85,30.31 62.69,30.06C62.31,29.38 62.58,26.87 63.51,22.52C64.43,18.17 65.59,13.75 66.96,9.27C68.34,4.76 69.44,1.96 70.27,0.87ZM49.2,8.17C49.75,7.62 51.11,7.75 53.28,8.55C55.49,9.32 56.91,10.22 57.55,11.24C57.94,11.75 58.13,12.18 58.13,12.54C58.13,12.89 57.95,13.56 57.6,14.55C57.12,15.96 56.32,17.32 55.2,18.63C54.15,20.01 53.41,21.05 52.99,21.75C52.93,21.88 52.85,22.02 52.75,22.18C52.66,22.31 52.55,22.44 52.42,22.57C52.32,22.66 52.21,22.74 52.08,22.81C51.99,22.84 51.89,22.86 51.79,22.86C50.32,22.86 50.72,21.34 52.99,18.3C54.66,16.12 55.51,14.6 55.54,13.74C55.6,13.13 55.51,12.68 55.25,12.39C55.03,12.07 54.43,11.67 53.47,11.19C52.03,10.46 50.85,10.09 49.92,10.09C49.22,10.09 48.79,9.91 48.63,9.56C48.47,9.21 48.66,8.74 49.2,8.17Z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M116.32,4.66C117.44,3.9 117.96,4.06 117.9,5.14C117.84,5.94 117.47,7.08 116.8,8.55C115.81,10.66 115.18,12.22 114.92,13.21L114.64,14.55L116.75,14.17C118.12,13.94 130.49,13.83 133.85,13.83C135.84,13.83 137.23,13.9 138.03,14.02C138.86,14.12 139.44,14.31 139.76,14.6C140.24,15.11 140.25,15.45 139.8,15.61C139.2,15.83 136.65,15.98 132.17,16.04C127.5,16.23 115,16.49 114.68,16.81C114.36,17.13 113.9,18.42 113.29,20.7C112.68,22.97 112.22,24.46 111.9,25.16C111.48,25.83 111.28,26.5 111.28,27.18C111.28,27.91 110.96,28.28 110.32,28.28C108.84,28.28 108.65,27.1 109.74,24.73C110.41,23.38 110.93,21.53 111.28,19.16C111.66,17.14 112.16,15.06 112.76,12.92C113.37,10.74 114,8.89 114.64,7.35C115.31,5.82 115.87,4.92 116.32,4.66ZM100.62,9.03C104.27,8.94 106.24,8.94 106.52,9.03C106.78,9.16 106.91,9.5 106.91,10.04C106.91,10.39 106.73,10.65 106.38,10.81C106.03,10.97 105.36,11.11 104.36,11.24C102.83,11.43 101.72,11.82 101.05,12.39C100.41,12.97 99.88,14.02 99.47,15.56C98.89,17.7 98.96,18.87 99.66,19.06C100.4,19.22 102.44,18.6 105.8,17.19C108.62,16.04 110.11,15.51 110.27,15.61C110.3,15.64 110.32,15.82 110.32,16.14C110.32,16.65 109.05,17.58 106.52,18.92C104,20.23 102.09,20.98 100.81,21.18C99.5,21.37 98.46,21.24 97.69,20.79C96.96,20.34 96.59,19.66 96.59,18.73C96.59,18.38 96.67,17.83 96.83,17.1C97.02,16.33 97.24,15.51 97.5,14.65C97.76,13.78 98.03,13.02 98.32,12.34C98.54,11.86 98.48,11.61 98.12,11.58C97.9,11.58 97.45,11.62 96.78,11.72C95.95,11.88 95.37,11.93 95.05,11.86C94.76,11.8 94.57,11.61 94.48,11.29C94.35,10.87 94.3,10.52 94.33,10.23C94.36,9.94 94.64,9.72 95.15,9.56C95.69,9.4 96.38,9.29 97.21,9.22C98.04,9.13 99.18,9.06 100.62,9.03Z"
+      android:fillColor="#ffffff"/>
+</vector>
