+++ app/src/main/res/layout/activity_join.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.user.JoinActivity">
+    tools:context=".presentation.user.JoinActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/tb_join"
