+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/main_blue"
-    tools:context=".ui.user.LoginActivity">
+    tools:context=".presentation.user.LoginActivity">
 
     <TextView
         android:id="@+id/txt_h1"
