+++ app/src/main/res/values-v31/themes.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <style name="Theme.YeoGida.Splash" parent="Theme.SplashScreen">
+        <item name="android:windowSplashScreenBackground">@color/main_blue</item>
+        <item name="android:windowSplashScreenAnimatedIcon">@drawable/ic_app</item>
+        <item name="android:windowSplashScreenAnimationDuration">1000</item>
+        <item name="postSplashScreenTheme">@style/Theme.YeoGida</item>
+        <item name="android:windowSplashScreenBrandingImage">@drawable/img_text_logo</item>
+    </style>
+</resources>
